requirements.txt
discord.py
requests
python-dotenv
flask
pytz
google-genai